pyproject.toml
[build-system]
requires = [
    "setuptools==65.5.1",
    "setuptools_scm[toml]==7.0.5",
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[project]
name = "prow-jobs-scraper"
description = "Scrape Prow for job results and export them to elasticsearch"
readme = "README.md"
keywords = ["prow"]
license = { file = "LICENSE" }
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "requests==2.28.2",
    "google-cloud-storage==2.8.0",
    "junitparser==2.8.0",
    "pydantic==1.10.7",
    "opensearch-py==2.2.0",
    "slack_sdk==3.20.2",
    "plotly==5.14.0",
    "kaleido==0.2.1",
    "python-dateutil==2.8.2",
]
dynamic = ["version"]

[project.scripts]
prow-jobs-scraper = "projects.prowjobsscraper.main:main"
jobs-auto-report = "projects.jobsautoreport.main:main"

[project.optional-dependencies]
test-runner = [
    "tox==4.4.8",
]
unit-tests = [
    "freezegun==1.2.2",
    "pytest==7.2.2",
    "pytest-cov==4.0.0",
    "pytest-httpserver==1.0.6",
]
coverage = [
    "coverage[toml]==7.2.2",
]
lint = [
    "black==23.3.0",
    "isort==5.12.0",
    "mypy==1.1.1",
]

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = clean,python3.9,report

[testenv]
usedevelop=True
commands = pytest --cov=src --cov-append --cov-report=term-missing --junitxml=junit-{envname}.xml --junit-prefix={envname} {posargs}
deps = .[unit-tests]
depends =
    {python3.9}: clean
    report: python3.9

[testenv:report]
deps = .[coverage]
skip_install = true
commands =
    coverage report
    coverage html
    coverage xml

[testenv:clean]
deps = .[coverage]
skip_install = true
commands = coverage erase
"""
